piniaに処理をリファクタリング
