+++ FRONTEND-ALCOHOL/package.json
@@ -12,6 +12,7 @@
   },
   "dependencies": {
     "axios": "^1.8.4",
+    "pinia": "^3.0.2",
     "vue": "^3.5.13",
     "vue-router": "^4.5.0"
   },
